UPDATES // Friends list LeftMenu.tsx and LeftMenu.css
